gwproxy/dns: Add DNS cache feature
Initial work to add DNS cache feature. The cache expiration is in
second unit, stored in struct gwp_dns_cfg::cache_expiry. If the
cache_expiry is set to zero or a negative value, the DNS cache
feature is disabled.

The user is expected to call `gwp_dns_cache_lookup()` to check and
get the cached result of a pair of hostname and service.

Signed-off-by: Ammar Faizi <ammarfaizi2@gnuweeb.org>