gwproxy: Swap out the old DNS setup for the new library version
The DNS library's all good and ready to go. Let's ditch the old DNS
system and plug in this new one. Building it as a library is awesome
because we are finally chopping the project into smaller, more
manageable and easier-to-handle bits.

Cramming everything into that single gwproxy.c file was a complete
headache. We should start doing things clean from now on...

Signed-off-by: Ammar Faizi <ammarfaizi2@gnuweeb.org>