io_uring: Introduce `GWP_CONN_FLAG_NO_CLOSE_FD` flag
The logging in `gwp_free_conn_pair()` when using io_uring shows all file
descriptors as -1. This is because when using io_uring, the client and
target file descriptors are closed asynchronously by the caller using
`io_uring_prep_close()` and the caller sets all fds to -1 before the
call to `gwp_free_conn_pair()` to avoid double-close.

Introduce `GWP_CONN_FLAG_NO_CLOSE_FD` to notify `gwp_free_conn_pair()`
that the file descriptors are owned by the caller so it's not allowed
to close them. Logging them are ok.

Signed-off-by: Ammar Faizi <ammarfaizi2@gnuweeb.org>