Makefile: Fix liburing build error
To build liburing, initially, the ./configure script needs to be
invoked. The default 'make' target when invoked in the liburing/
directory will invoke the ./configure script if it's needed.

Don't run the 'make' command in the liburing/src/ directory. Run it in
liburing/ directory to cover the ./configure script invokation.

Running the make command in liburing/src/ only works when the ./configure
script has been invoked first. If it has not been invoked, the build will
lead to this error:

  $ make -j
  make[1]: Entering directory '/home/ammarfaizi2/p/gwproxy/src/liburing/src'
  Makefile:37: ../config-host.mak: No such file or directory
  make[1]: *** No rule to make target '../config-host.mak'.  Stop.
  make[1]: Leaving directory '/home/ammarfaizi2/p/gwproxy/src/liburing/src'
  make: *** [Makefile:75: src/liburing/src/liburing.a] Error 2

Fixes: 10ee5b83b50f ("io_uring: Add liburing to the code base")
Signed-off-by: Ammar Faizi <ammarfaizi2@gnuweeb.org>