gwproxy: Integrate new experimental raw DNS backend to the code base
The raw DNS backend is marked as experimental and disabled by default.
It can be enabled at build time with the configure option --use-raw-dns,
and at runtime with the gwproxy option -r 1.

When experimental raw DNS is enabled in gwproxy, the program benefits from:
- disabling DNS worker threads
- disabling eventfd that used to communicate between threads
- disabling condition variables and mutex locks
- disabling reference counting (no need to prevent UAF)

Use fallback mechanism for raw DNS
If *_PREFER_* is used in the restyp, the raw DNS backend will attempt to
retry DNS query with different address family if it failed.

Add new cmdline arguments:
- raw-dns: enables the experimental raw DNS feature
- dns-server: sets the default DNS server
- session-map-cap: sets the default capacity for the session_map and stack.arr arrays

The struct gwp_dns_ctx is now exported due to the need for accessing it
outside dns.c.

Signed-off-by: Ahmad Gani <reyuki@gnuweeb.org>