epoll: Improve log readability and efficiency
I noticed from the info log that the connection is reported as closed
twice, even though I only tested with a single curl. After investigating,
I realized it's not an error, just a misleading log message.

This change fixes the confusing log and also improves efficiency:
by moving the accept4 syscall to the top of __handle_ev_accept, we can
eliminate an unnecessary allocation.

Signed-off-by: Ahmad Gani <reyuki@gnuweeb.org>
Link: https://lore.gnuweeb.org/gwml/20250817050551.92428-1-reyuki@gnuweeb.org
Signed-off-by: Ammar Faizi <ammarfaizi2@gnuweeb.org>